add verification thread & comment

+++ src/Applications/use_case/DeleteCommentUseCase.js
@@ -1,3 +1,4 @@
+const AuthorizationError = require('../../Commons/exceptions/AuthorizationError');
 const CommentRepository = require('../../Domains/comments/CommentRepository');
 
 class DeleteCommentUseCase {
@@ -5,17 +6,28 @@ class DeleteCommentUseCase {
      *
      * @param {Object} dependencies - dependencies of use case
      * @param {CommentRepository} dependencies.commentRepository - comment repository
+     * @param {ThreadRepository} dependencies.threadRepository - thread repository
      */
-    constructor({ commentRepository }) {
+    constructor({ commentRepository, threadRepository }) {
         this._commentRepository = commentRepository;
+        this._threadRepository = threadRepository;
     }
 
     /**
      * @param {string} id
      * @returns {Promise<void>}
      * @throws {NotFoundError}
+     * @throws {AuthorizationError}
      */
-    async execute(id) {
+    async execute(id, owner) {
+        const result =
+            await this._commentRepository.verifyCommentExists(id);
+        if (result.owner !== owner) {
+            throw new AuthorizationError('anda tidak memiliki akses');
+        }
+        await this._threadRepository.verifyThreadExists(
+            result.thread,
+        );
         await this._commentRepository.deleteComment(id);
     }
 }
